requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anthropic==0.84.0
anyio==4.12.1
asgiref==3.11.1
billiard==4.2.4
boto3==1.42.63
botocore==1.42.63
celery==5.6.2
certifi==2026.2.25
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
distro==1.9.0
Django==6.0.3
django-cors-headers==4.9.0
django-storages==1.14.6
djangorestframework==3.16.1
docstring_parser==0.17.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.13.0
jmespath==1.1.0
kombu==5.6.2
packaging==26.0
pillow==12.1.1
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
pydantic==2.12.5
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
redis==7.3.0
s3transfer==0.16.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.5
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
vine==5.1.0
wcwidth==0.6.0
